Separate FormParams class
